feat(search): process [C-h] and [C-?] as representations of backspace (#1857)
In the conventional terminal protocol, Backspace can be transmitted as
the code \x08 or \x7F depending on the terminal.  Ctrl+Backspace can
also be transmitted as the code \x08 or \x7F.  These overlap with the
code for Ctrl+H and Ctrl+?.  The crossterm library does not try to
handle these terminal dependencies (probably because it is hard to
resolve it perfectly).  To provide a consistent experience among
terminals, we assign to C-h and C-? the same feature as backspace.

Note: The crossterm seems to produce Ctrl+Backspace only in the
extended keyboard protocol, so we can trust crossterm particularly for
Ctrl+Backspace.  For this reason, we keep the feature of removing a
backward word by Ctrl+Backspace.

https://github.com/atuinsh/atuin/issues/1753