Added "accept-to-edit" for (Back)Space / Home / Left
The above keys edit the selected command (as with hitting Tab), but
like if using them in the shell. They only trigger when the search
input is currently empty, and must be enabled individually with
the respective boolean config options:

* exit_with_backspace: right-trim command and remove last character

* exit_with_space: right-trim and append a space character

* exit_with_home: start editing at position 0

* exit_with_cursor_left: right-trim and start editing at last character

* exit_positions_cursor: enable cursor positioning for Home and Left
  keys (bash-only)

Updated bash init to support cursor positioning. For other shells, Home
and Left keys will behave like Tab.

Resolves #1906