chore: update rust to `1.80.1` (#2362)
fixes #2360

https://blog.rust-lang.org/2024/08/08/Rust-1.80.1.html

`1.80.1` fixes a fairly nasty miscompilation, but that's not my motivation for opening this PR.

If I'm understanding correctly, because the rust patch version was not specified in `rust-toolchain.toml`, nix automatically upgraded to the new version. This broke the hash in `flake.nix` causing the flake to fail to build. Making this upgrade to `1.80.1` explicit, fixes this issue.