Bind Ctrl-v to visually editing the selected command
My old workflow was:
  - select the command I want
  - hit tab, so now it is on the command line, atuin is gone, nothing has been executed
  - I use Vim mode on the command line, so Escape puts me in normal mode
  - vv launches my favorite editor
  - edit, write, and quit which makes the updated command execute

My new workflow follows in the footsteps of fc:
  - select the command I want
  - hit Ctrl-v, launches my favorite editor on the text of that command
  - edit, write, and quit which makes the updated command execute

I tried to make this as much like fc as I could because I feel like atuin
hides and replaces fc.  Emacs-mode on the command line requires two control
keys to jump into the editor.  Vim-mode needs you to get into normal mode,
that is NOT inserting, then hit v.  Everything else in atuin is a single
key, and when stepping through the history list, you're already not in
insert mode.  Like fc, I look for editors in this order:
  - the value of $FCEDIT (I know this feels weird, but you are "fixing a command")
  - the value of $EDITOR
  - the value of $VISUAL
  - or else vim