feat: add various acceptance keys
This change introduces (optional) acceptance keys of Backspace and Left
Arrow, when at the start of a line. These two are common muscle memory
actions for users.

The configuration defaults to false so as not to disrupt existing user
patterns, and implements #[serde(default)] to avoid breaking existing
configurations where these options don't exist.