ci: add Windows builds, second try (#2966)
Hi @ellie, I don't want to be a burden, but I'd like to ask if you could
please reconsider adding Windows builds now that PowerShell support is
merged (#2543). 😇

Most people use PowerShell on Windows, so they'd need an artifact in the
future releases in order to use it. Otherwise, they'd have to install
Rust and compile it themselves, which is not user-friendly.

I'd like to point out that I didn't have issues with building Atuin on
Windows when I was rebasing the PowerShell PR (except some initial stuff
like `#[cfg]` or a few warnings which were fixed in #2856), so I don't
expect this to introduce many problems.

This is essentially an updated version of #2714.

## Checks
- [x] I am happy for maintainers to push small adjustments to this PR,
to speed up the review cycle
- [x] I have checked that there are no existing pull requests for the
same thing