fix(bash): fix issues with intermediate key sequences in the vi editing mode (#2977)
In #2953, I chose `\e[0;<n>A` as special key sequences for
`enter_accept` in Bash, but I realized that the sequence `\e[0;<n>A`
doesn't actually work in the `vi-insert` keymap. I instead suggest using
the key sequences of the form `\C-x\C-_A0\a` (99ddcbf7), which is
unlikely to conflict with the users' keybindings. In addition, for Bash
<= 3.2, I also noticed that we cannot insert `$READLINE_LINE` in the
command line when we are in the `vi-command` keymap. I suggest switching
from `vi-command` to `vi-insert` temporarily (e6c4e9fd).

## Checks
- [x] I am happy for maintainers to push small adjustments to this PR,
to speed up the review cycle
- [x] I have checked that there are no existing pull requests for the
same thing